# optimize some misc runtime functions