For example, `datetime.date() == None` was slow since date_richcompare would call PyObject_HasAttrString which was slow if the object didn't have that attribute.

I think the perf changes are mostly spurious:
```
      django_template3.py             4.9s (2)             4.9s (2)  +0.6%
            pyxl_bench.py             3.1s (2)             3.0s (2)  -2.2%
sqlalchemy_imperative2.py             3.5s (2)             3.4s (2)  -1.2%
                  geomean                 3.7s                 3.7s  -1.0%
```
The optimizations were derived from django_template3 :P